+++ src/PNGChunk.cpp
@@ -8,9 +8,9 @@
 //
 // ************************************************************************** //
 std::vector<std::byte>
-readPNGHeader( std::string filename )
+readPNGSignature( const std::string &filename )
 {
-    std::vector<std::byte> headerData;
+    std::vector<std::byte> sigData;
     char byte;
     std::ifstream pngFile( filename, std::ios::binary );
     
@@ -21,33 +21,40 @@ readPNGHeader( std::string filename )
         for( size_t i = 0; i < pngFileSignatureSize; i++ )
         {
             pngFile.read(&byte, 1);
-            headerData.push_back(std::byte(byte));
+            sigData.push_back(std::byte(byte));
         }
     }
 
     pngFile.close();
 
-    return( headerData );
+    return( sigData );
 }
 
 // ************************************************************************** //
-//
-// filename - file to read data from
-// offset - number of bytes to skip before starting reading.
-// bytesRead - return number of bytes this chunk takes up.
-// 
-// Becuase PNG's are stored in network byte order, we should call ntoh[ls]
+// \brief Reads a png file's chunk data at location offset.
+//        Returns the number of bytes 
+// Because PNG's are stored in network byte order, we should call ntoh[ls]
 // as we read in the data to get the correct endianess for our machine.
+// will return a chunk of size bytes 0 on error.
 //
+// \param[in]  filename  - name of the png file to read in.
+// \param[in]  offset    - how many bytes into the file to read in the png chunkd.
+//
+// \return PNGChunk
 // ************************************************************************** //
 PNGChunk 
-readPNGChunk( std::string filename, size_t offset, size_t &bytesRead )
+readPNGChunk( const std::string &filename, size_t offset )
 {
 
     PNGChunk chunk;
 
     // open the file for reading.
     std::ifstream pngFile( filename, std::ios::binary );
+    
+    uint32_t length(0);
+    uint32_t typeCode(0);
+    std::vector<std::byte> data(0);
+    uint32_t crc(0);
 
     if( pngFile.good() )
     {
@@ -55,21 +62,23 @@ readPNGChunk( std::string filename, size_t offset, size_t &bytesRead )
         pngFile.ignore(offset);
 
         // Read the length element, this will tell how many bytes to read from the data portion of the chunk.
-        pngFile.read(reinterpret_cast<char*>(&chunk.length), sizeof(chunk.length));
-        chunk.length = ntohl(chunk.length);
+        pngFile.read(reinterpret_cast<char*>(&length), sizeof(length));
+        chunk.setLength(ntohl(length));
 
         // Next read the typeCode, this will tell us what kind of data this chunk contains.
-        pngFile.read(reinterpret_cast<char*>(&chunk.typeCode), sizeof(chunk.typeCode));
-        chunk.typeCode = ntohl(chunk.typeCode);
+        pngFile.read(reinterpret_cast<char*>(&typeCode), sizeof(typeCode));
+        chunk.setTypeCode(ntohl(typeCode));
 
         // Resize the data to the length read in.
-        chunk.data.resize(chunk.length);
+        data.resize(chunk.getLength());
 
         // Loop over the bytes and read them in.
-        pngFile.read(reinterpret_cast<char*>(&chunk.data[0]), chunk.length);
+        pngFile.read(reinterpret_cast<char*>(&data[0]), chunk.getLength());
+        chunk.setData(data);
     
         // Lastly read in the CRC. We won't do error checking on the chunks quite yet.
-        pngFile.read(reinterpret_cast<char*>(&chunk.crc), sizeof(chunk.crc));
+        pngFile.read(reinterpret_cast<char*>(&crc), sizeof(crc));
+        chunk.setCRC(ntohl(crc));
 
         std::cout << __FILE__ << ":" << __LINE__ << ":";
         std::cout << "ADD PNG CRC CHECK HERE" << std::endl;
@@ -78,11 +87,21 @@ readPNGChunk( std::string filename, size_t offset, size_t &bytesRead )
 
     pngFile.close();
 
-    // The size of all the data elements in the chunk is the number of bytes read.
-    // You can then use this to compute the offset for the next chunk.
-    bytesRead = sizeof(chunk.length) + sizeof(chunk.typeCode) + chunk.data.size() * sizeof(std::byte) + sizeof(chunk.crc);
+
 
     return( chunk );
+}
+
+size_t
+PNGChunk::getSizeInBytes()
+{
+    size_t chunkSize(0);
+
+    // The size of all the data elements in the chunk is the number of bytes read.
+    // You can then use this to compute the offset for the next chunk.
+    chunkSize = sizeof(this->_length) + sizeof(this->_typeCode) + this->_data.size() * sizeof(std::byte) + sizeof(this->_crc);
+
+    return(chunkSize);
 }
 
 std::string PNGChunk::toString()
@@ -93,31 +112,31 @@ std::string PNGChunk::toString()
        << "Start of Chunk" << std::endl << std::endl;
 
 
-    ss << "length:   " << this->length   << std::endl;
-    ss << "typeCode: " << this->typeCode << std::endl;
+    ss << "length:   " << this->getLength()   << std::endl;
+    ss << "typeCode: " << this->getTypeCode() << std::endl;
 
     ss << "data: " << std::endl;
 
     ss << std::hex << std::uppercase << std::setfill('0'); 
-    for (size_t i = 0; i < this->data.size(); i++)
+    for (size_t i = 0; i < this->getData().size(); i++)
     {
 
-        if( i % 2 == 0 )
+        if( i % 1 == 0 )
         {
-            ss << " ";
+            ss << "0x";
             if( i % 16 == 0 )
             {
                 ss << std::endl;
             }
         }
         // Cast std::byte to uint32_t and print as hex with 2 digits
-        ss << std::setw(2) << static_cast<uint32_t>(this->data[i]);
+        ss << std::setw(2) << static_cast<uint32_t>(this->getData()[i]) << ", ";
     }
 
 
     ss << std::dec << std::nouppercase << std::endl;
 
-    ss << "CRC:      " << this->crc << std::endl << std::endl;
+    ss << "CRC:      " << this->getCRC() << std::endl << std::endl;
 
     ss << "End of Chunk" << std::endl
        << "-----------------------------------------------------";